+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Change Log
 ==========
+            
+## next
+
+* `TraktV2`: add `isUnauthorized(response)`, `isAccountLocked(response)` and `isNotVip(response)` helper methods.
 
 ## 6.15.0
 _2024-08-30_